refactor(utils): merges create-reducer and create-constants into a single utils/index.js file; removes eslint new-caps rule which doesnt play nicely with Immutable.js
