--- src/utils/create-constants.js
@@ -1,9 +1,0 @@
-import assign from 'object-assign';
-
-export default function createConstants (...constants) {
-  return constants.reduce((acc, constant) => {
-    return assign(acc, {
-      [constant] : constant
-    });
-  }, {});
-}
