--- src/utils/create-reducer.js
@@ -1,7 +1,0 @@
-export default function createReducer (initialState, reducerMap) {
-  return (state = initialState, action) => {
-    const reducer = reducerMap[action.type];
-
-    return reducer ? reducer(state, action.payload) : state;
-  };
-}
